move SearchBar to legacy components
