+++ src/components/legacy/SearchBar/SearchBar.js
