+++ src/components/legacy/SearchBar/SearchInput.js
