+++ src/components/legacy/SearchBar/package.json
