+++ storybook/storyLoader.js
@@ -28,7 +28,7 @@ function loadStories() {
 	require('../src/components/RetryGraphQueryBox/RetryGraphQueryBox.stories');
 	require('../src/components/ReviewSummary/ReviewSummary.stories');
 	require('../src/components/SWActivationCard/SWActivationCard.stories');
-	require('../src/components/SearchBar/SearchBar.stories');
+	require('../src/components/legacy/SearchBar/SearchBar.stories');
 	require('../src/components/Table/Table.stories');
 	require('../src/components/TextInput/TextInput.stories');
 	require('../src/components/Toast/ToastCard.stories');
@@ -62,7 +62,7 @@ const stories = [
 	'../src/components/RetryGraphQueryBox/RetryGraphQueryBox.stories',
 	'../src/components/ReviewSummary/ReviewSummary.stories',
 	'../src/components/SWActivationCard/SWActivationCard.stories',
-	'../src/components/SearchBar/SearchBar.stories',
+	'../src/components/legacy/SearchBar/SearchBar.stories',
 	'../src/components/Table/Table.stories',
 	'../src/components/TextInput/TextInput.stories',
 	'../src/components/Toast/ToastCard.stories',
